+++ include/seqan3/argument_parser/argument_parser.hpp
@@ -241,7 +241,7 @@ public:
                     char const short_id,
                     std::string const & long_id,
                     std::string const & desc,
-                    option_spec const spec = option_spec::defaulted,
+                    option_spec const spec = option_spec::standard,
                     validator_type option_validator = validator_type{}) // copy to bind rvalues
     {
         if (sub_parser != nullptr)
@@ -266,7 +266,7 @@ public:
                   char const short_id,
                   std::string const & long_id,
                   std::string const & desc,
-                  option_spec const spec = option_spec::defaulted)
+                  option_spec const spec = option_spec::standard)
     {
         verify_identifiers(short_id, long_id);
         // copy variables into the lambda because the calls are pushed to a stack
@@ -488,22 +488,22 @@ public:
 
     /*!\brief Adds an help page section to the seqan3::argument_parser.
      * \param[in] title The title of the section.
-     * \param[in] spec Whether to always display this section title (seqan3::option_spec::defaulted), only when showing
+     * \param[in] spec Whether to always display this section title (seqan3::option_spec::standard), only when showing
      *                 the advanced help page (seqan3::option_spec::advanced) or never (seqan3::option_spec::hidden).
      * \details This only affects the help page and other output formats.
      */
-    void add_section(std::string const & title, option_spec const spec = option_spec::defaulted)
+    void add_section(std::string const & title, option_spec const spec = option_spec::standard)
     {
         std::visit([&] (auto & f) { f.add_section(title, spec); }, format);
     }
 
     /*!\brief Adds an help page subsection to the seqan3::argument_parser.
      * \param[in] title The title of the subsection.
-     * \param[in] spec Whether to always display this subsection title (seqan3::option_spec::defaulted), only when showing
+     * \param[in] spec Whether to always display this subsection title (seqan3::option_spec::standard), only when showing
      *                 the advanced help page (seqan3::option_spec::advanced) or never (seqan3::option_spec::hidden).
      * \details This only affects the help page and other output formats.
      */
-    void add_subsection(std::string const & title, option_spec const spec = option_spec::defaulted)
+    void add_subsection(std::string const & title, option_spec const spec = option_spec::standard)
     {
         std::visit([&] (auto & f) { f.add_subsection(title, spec); }, format);
     }
@@ -511,13 +511,13 @@ public:
     /*!\brief Adds an help page text line to the seqan3::argument_parser.
      * \param[in] text The text to print.
      * \param[in] is_paragraph Whether to insert as paragraph or just a line (Default: false).
-     * \param[in] spec Whether to always display this line (seqan3::option_spec::defaulted), only when showing
+     * \param[in] spec Whether to always display this line (seqan3::option_spec::standard), only when showing
      *                 the advanced help page (seqan3::option_spec::advanced) or never (seqan3::option_spec::hidden).
      * \details
      * If the line is not a paragraph (false), only one line break is appended, otherwise two line breaks are appended.
      * This only affects the help page and other output formats.
      */
-    void add_line(std::string const & text, bool is_paragraph = false, option_spec const spec = option_spec::defaulted)
+    void add_line(std::string const & text, bool is_paragraph = false, option_spec const spec = option_spec::standard)
     {
         std::visit([&] (auto & f) { f.add_line(text, is_paragraph, spec); }, format);
     }
@@ -525,7 +525,7 @@ public:
     /*!\brief Adds an help page list item (key-value) to the seqan3::argument_parser.
      * \param[in] key  The key of the key-value pair of the list item.
      * \param[in] desc The value of the key-value pair of the list item.
-     * \param[in] spec Whether to always display this list item (seqan3::option_spec::defaulted), only when showing
+     * \param[in] spec Whether to always display this list item (seqan3::option_spec::standard), only when showing
      *                 the advanced help page (seqan3::option_spec::advanced) or never (seqan3::option_spec::hidden).
      *
      * \details
@@ -542,7 +542,7 @@ public:
      */
     void add_list_item(std::string const & key,
                        std::string const & desc,
-                       option_spec const spec = option_spec::defaulted)
+                       option_spec const spec = option_spec::standard)
     {
         std::visit([&] (auto & f) { f.add_list_item(key, desc, spec); }, format);
     }